requirements.txt
streamlit==1.51.0
pandas==2.3.3
numpy==2.3.4
openpyxl==3.1.5
python-docx==1.2.0
PyPDF2==3.0.1
pdfplumber==0.11.8
fpdf==1.7.2
plotly==6.4.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
